refactor: relocate shared RL server/worker/utils into arctic_platform.common
Move the protocol-agnostic training infrastructure (DeepSpeed worker, HTTP/Ray
servers, ray cluster helpers, and low-level utils) out of arctic_platform.rl
into a new arctic_platform.common package, leaving back-compat shims at the old
arctic_platform.rl.* import paths.

Behavior-preserving move: every common/ module is byte-identical to its former
rl/ counterpart except for rewriting imports of the relocated siblings
(arctic_platform.rl.{deepspeed_worker,http_server,ray_server,ray_cluster,server,
utils} -> arctic_platform.common.*). RL-specific modules (processors,
zorro_train, client, config, weight_sync) stay in rl/ and are still imported
from there. No new functionality; this sets up shared code for the upcoming
SFT integration.
