+++ arctic_platform/rl/ray_cluster.py
@@ -13,251 +13,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-# Copyright 2025 Snowflake Inc.
-# SPDX-License-Identifier: Apache-2.0
+"""Back-compat shim — prefer ``arctic_platform.common.ray_cluster``."""
 
-"""Multinode Ray cluster bootstrap.
-
-Behaviour:
-1. If a Ray cluster is already running, attach to it.
-2. Otherwise start a head node locally and, if ``/job/hostfile`` lists remote
-   hosts, fan out ``ray start`` via ``pdsh``.
-3. On exit, only tear down a cluster this process started.
-"""
-
-from __future__ import annotations
-
-import atexit
-import logging
-import os
-import shutil
-import socket
-import subprocess
-import sys
-import tempfile
-from pathlib import Path
-
-import ray
-from ray._private.utils import read_ray_address
-
-from arctic_platform.rl.utils.debug import pr0
-
-logger = logging.getLogger(__name__)
-
-_HOSTFILE = "/job/hostfile"
-_DEFAULT_RAY_PORT = 6379
-_DEFAULT_RAY_DASHBOARD_PORT = 8265
-
-_spawned_cluster = False
-_spawned_temp_dir: str | None = None
-
-
-def _is_port_free(port: int) -> bool:
-    with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as s:
-        try:
-            s.bind(("", port))
-        except OSError:
-            return False
-    return True
-
-
-def _free_port() -> int:
-    with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as s:
-        s.bind(("", 0))
-        return s.getsockname()[1]
-
-
-def _resolve_port(env_var: str, default: int) -> int:
-    """Resolve a Ray port.
-
-    If ``env_var`` is set, the user-supplied port is returned verbatim (no
-    fallback: a subsequent bind collision is surfaced as a hard failure, so the
-    user's explicit choice is never silently overridden). Otherwise return
-    ``default`` if free, falling back to an OS-assigned free port.
-    """
-    override = os.environ.get(env_var)
-    if override is not None:
-        return int(override)
-    if _is_port_free(default):
-        return default
-    port = _free_port()
-    logger.info(f"Port {default} is in use for {env_var}; falling back to free port {port}")
-    return port
-
-
-def primary_ip() -> str:
-    """Return the routable IP of this machine."""
-    s = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
-    try:
-        s.connect(("10.255.255.255", 1))
-        return s.getsockname()[0]
-    except OSError:
-        return "127.0.0.1"
-    finally:
-        s.close()
-
-
-def _peer_hosts() -> list[str]:
-    """Return remote host entries from ``/job/hostfile``, excluding this machine."""
-    try:
-        with open(_HOSTFILE, encoding="utf-8") as f:
-            lines = f.readlines()
-    except OSError:
-        return []
-    head = primary_ip()
-    hosts = [ln.split()[0] for ln in lines if ln.strip() and not ln.strip().startswith("#")]
-    return [h for h in hosts if h != head]
-
-
-def _pdsh(hosts: list[str], cmd: list[str], **kwargs) -> subprocess.CompletedProcess:
-    env = {**os.environ, "PDSH_RCMD_TYPE": "ssh"}
-    return subprocess.run(["pdsh", "-w", ",".join(hosts)] + cmd, env=env, **kwargs)
-
-
-def _ray_bin() -> str:
-    return str(Path(sys.executable).resolve().parent / "ray")
-
-
-def init_ray_cluster(auto_attach: bool = True) -> None:
-    """Attach to an existing Ray cluster or start one (with multinode support).
-
-    Args:
-        auto_attach: If True (default), attempt to attach to a pre-existing Ray
-            cluster before starting a new one. The attach is only honored if the
-            existing cluster exposes GPU resources; CPU-only clusters (e.g. one
-            started by an unrelated integration that also uses Ray) are ignored
-            and a new cluster is started instead. Set to False to always start a
-            fresh cluster.
-    """
-    global _spawned_cluster, _spawned_temp_dir
-
-    # 1. Try attaching to a running cluster (only if it has GPUs).
-    if auto_attach:
-        try:
-            ray.init(address="auto", ignore_reinit_error=True, log_to_driver=True)
-            gpus = ray.cluster_resources().get("GPU", 0)
-            if gpus > 0:
-                logger.info(
-                    "Attached to existing Ray cluster at %s (%.0f GPU(s))",
-                    ray.get_runtime_context().gcs_address,
-                    gpus,
-                )
-                return
-            logger.info("Existing Ray cluster has no GPU resources; ignoring and starting a new cluster.")
-            ray.shutdown()
-        except ConnectionError:
-            pass
-
-    # 2. Start a new head node. Pick non-colliding ports so we don't clash with
-    # any pre-existing (e.g. CPU-only) cluster on this host. Use a unique
-    # ``--temp-dir`` so ``_shutdown()`` can stop only our processes by matching
-    # that temp-dir basename on their command lines.
-    r = _ray_bin()
-    ray_port = _resolve_port("RAY_PORT", _DEFAULT_RAY_PORT)
-    dashboard_port = _resolve_port("RAY_DASHBOARD_PORT", _DEFAULT_RAY_DASHBOARD_PORT)
-    # Honor a pre-created temp dir if supplied (the test harness pre-creates a unique one per client session so it
-    # can reap exactly this cluster by basename without racing parallel workers); otherwise pick our own. Must stay
-    # shallow -- Ray's AF_UNIX socket paths under <temp-dir>/session_*/sockets/ cannot exceed 107 bytes.
-    _spawned_temp_dir = os.environ.get("ARL_RAY_TEMP_DIR") or tempfile.mkdtemp(prefix="ray_arctic_")
-    start_cmd = [
-        r,
-        "start",
-        "--head",
-        f"--port={ray_port}",
-        f"--dashboard-port={dashboard_port}",
-        f"--temp-dir={_spawned_temp_dir}",
-        "--disable-usage-stats",
-    ]
-    # Honor a pre-set worker-port range if supplied. Ray assigns its CoreWorker gRPC ports from a fixed default
-    # range (10002+) shared by every cluster on the host, so two concurrent per-worker clusters (the test harness's
-    # partitioned pytest-xdist path) collide on a worker port -- a fatal, non-retried bind error that aborts the
-    # CoreWorker. The harness strides this range per worker to keep concurrent clusters disjoint. Pure opt-in; unset
-    # in production, where Ray's default range applies.
-    min_worker_port = os.environ.get("ARL_RAY_MIN_WORKER_PORT")
-    max_worker_port = os.environ.get("ARL_RAY_MAX_WORKER_PORT")
-    if min_worker_port and max_worker_port:
-        start_cmd += [f"--min-worker-port={min_worker_port}", f"--max-worker-port={max_worker_port}"]
-    subprocess.run(
-        start_cmd,
-        check=True,
-        timeout=300,
-        env=os.environ,
-    )
-    pr0(f"[init_ray_cluster] ray started with port {ray_port} and dashboard port {dashboard_port}")
-
-    # 3. Start workers on peer nodes (if any).
-    peers = _peer_hosts()
-    pr0(f"[init_ray_cluster] peers: {peers}")
-    gcs = read_ray_address(_spawned_temp_dir)
-    if peers:
-        logger.info("Starting Ray workers on %s (address=%s)", peers, gcs)
-        result = _pdsh(
-            peers,
-            [r, "start", f"--address={gcs}", f"--temp-dir={_spawned_temp_dir}", "--disable-usage-stats"],
-            check=False,
-            timeout=600,
-        )
-        if result.returncode != 0:
-            logger.warning("pdsh ray start returned exit code %d", result.returncode)
-
-    _spawned_cluster = True
-    # Use the explicit GCS address instead of ``auto`` so we don't accidentally
-    # reconnect to a different Ray cluster running on this host.
-    ray.init(address=gcs, ignore_reinit_error=True, log_to_driver=True)
-    pr0(f"[init_ray_cluster] ray initialized with address {gcs}")
-    resources = ray.available_resources()
-    logger.info(
-        "Ray cluster: %.0f GPU(s), %.0f CPU(s), %d node(s)",
-        resources.get("GPU", 0),
-        resources.get("CPU", 0),
-        sum(1 for k in resources if k.startswith("node:")),
-    )
-
-
-def _reset_cached_ray_address() -> None:
-    """Drop Ray's cached default cluster address.
-
-    After we kill the cluster we started, a later ``ray.init(address="auto")``
-    must not resolve to (and then hard-abort on) the now-dead cluster.
-    """
-    for mod in ("ray._common.utils", "ray._private.utils"):
-        try:
-            __import__(mod, fromlist=["reset_ray_address"]).reset_ray_address()
-            break
-        except Exception:
-            continue
-    os.environ.pop("RAY_ADDRESS", None)
-
-
-def _shutdown() -> None:
-    global _spawned_cluster, _spawned_temp_dir
-    try:
-        ray.shutdown()
-    except Exception:
-        pass
-    if not _spawned_cluster:
-        return
-    _spawned_cluster = False
-    if _spawned_temp_dir is None:
-        return
-
-    # ``ray.shutdown()`` only disconnects this driver; the ``ray start --head``
-    # daemon (gcs_server / raylet / dashboard / monitor / log_monitor) keeps
-    # running. ``ray stop`` would kill it, but it is *global* -- it kills every
-    # Ray process on the host by name, including any coexisting cluster owned by
-    # this user -- and it has no ``--temp-dir`` scoping option (that flag belongs
-    # to ``ray start``). So we surgically SIGKILL only the processes belonging to
-    # the cluster we started, matched by the unique temp-dir basename that
-    # appears on each daemon's command line (raylet/gcs_server/dashboard/...).
-    basename = os.path.basename(_spawned_temp_dir)
-    kill_cmd = ["pkill", "-9", "-f", basename]
-    subprocess.run(kill_cmd, check=False, timeout=120)
-    peers = _peer_hosts()
-    if peers:
-        _pdsh(peers, kill_cmd, check=False, timeout=180, capture_output=True)
-    shutil.rmtree(_spawned_temp_dir, ignore_errors=True)
-    _spawned_temp_dir = None
-    _reset_cached_ray_address()
-
-
-atexit.register(_shutdown)
+from arctic_platform.common.ray_cluster import *  # noqa: F401,F403
