+++ arctic_platform/rl/utils/__init__.py
@@ -13,52 +13,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .batch import combine_metric_microbatches
-from .batch import combine_metric_shards
-from .batch import http_split_batch
-from .batch import log_dp_shard_tokens
-from .batch import merge_dict_shards
-from .batch import ray_split_batch
-from .batch import shard_token_stats
-from .batch import split_dict
-from .batch import unpack_batch
-from .cuda_ipc import merge_cuda_ipc_payloads
-from .debug import ProfilerContext
-from .debug import SynchronizedWallClockTimerSimple
-from .record_replay import record_replay_generation
-from .server_models import GenerateRequest
-from .server_models import JobConfig
-from .server_models import LogProbsRequest
-from .server_models import OperationRequest
-from .server_models import ResetPrefixCacheRequest
-from .server_models import SaveRequest
-from .server_models import StepRequest
-from .server_models import WeightNormRequest
-from .server_models import WeightSyncRequest
-from .server_models import build_model_config
+"""Back-compat shim — utils live in ``arctic_platform.common.utils``."""
 
-__all__ = [
-    "unpack_batch",
-    "merge_dict_shards",
-    "combine_metric_shards",
-    "combine_metric_microbatches",
-    "split_dict",
-    "http_split_batch",
-    "ray_split_batch",
-    "shard_token_stats",
-    "log_dp_shard_tokens",
-    "merge_cuda_ipc_payloads",
-    "ProfilerContext",
-    "record_replay_generation",
-    "SynchronizedWallClockTimerSimple",
-    "JobConfig",
-    "GenerateRequest",
-    "LogProbsRequest",
-    "StepRequest",
-    "SaveRequest",
-    "ResetPrefixCacheRequest",
-    "OperationRequest",
-    "WeightSyncRequest",
-    "WeightNormRequest",
-    "build_model_config",
-]
+from arctic_platform.common.utils import *  # noqa: F401,F403
+from arctic_platform.common.utils import __all__  # noqa: F401
